refactor(extract/paloalto): move PAN-OS interpretation to internal/panos
Extract the changes[] interpretation (Stanza/Change/Interval +
StanzaIntervals + parse helpers + transition priority) into a shared
internal/panos package so the upcoming list extractor can reuse it
instead of duplicating the logic. The json extractor now builds
panos.Stanza and calls panos.StanzaIntervals; the spec unit test moves to
the panos package. Output is unchanged (identical golden). As internal/,
these types are a legitimate cross-package API rather than test-only
exports.
