--- pkg/extract/paloalto/json/export_test.go
@@ -1,6 +1,0 @@
-package json
-
-// PanosStanzaIntervals exposes panosStanzaIntervals for the white-box test in
-// package json_test. The PanosStanza / PanosChange / PanosInterval value types
-// it operates on are exported in json.go.
-var PanosStanzaIntervals = panosStanzaIntervals
